# [v7r2] Optimise ClassAd.asJDL