@fstagni 2021-06-07
This looks OK to me. There is no specific unit test for this module, if you feel like it, add one (which may help discover other optimizations...)