@chrisburr 2021-06-07
For now I'd rather add it to my todo list as I think the sandbox store and CS issues are more pressing.

(For anyone non-LHCb reading, these are mostly LHCb specific issues and definitely not v7r2 related)